% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmm_cols.R
\name{jmm_cols}
\alias{jmm_cols}
\title{Corporate JMM colors}
\usage{
jmm_cols(...)
}
\arguments{
\item{...}{Corporate colors}
}
\value{
The hex code
}
\description{
This function extracts JMM colors as hex codes
}
\examples{
\dontrun{

jmm_colors <- c(
`dark blue  = "#123549",
`green      = "#7AD651",
`blue       = "#2f97c4",
`yellow     = "#f1c600",
`light grey = "#D9D9D9",
`dark grey  = "#A3A3A3")

jmm_cols("blue")
}
}
