% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmm_pal.R
\name{jmm_pal}
\alias{jmm_pal}
\title{Function to interpolate a jmm color palette (jmm_pal)}
\usage{
jmm_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{the corporate color palette to use}

\item{n}{the number of colors the color palette should have}
}
\value{
The n hex code colors
}
\description{
This function interpolates a jmm color palette to have n colors
}
\examples{
\dontrun{

jmm_palettes <- list(
`main  = jmm_cols("blue", "green", "yellow"),

`cool  = jmm_cols("dark blue", "blue"),

`cool grey  = jmm_cols("blue", "light grey"),

`grey blue = jmm_cols("light grey", "dark blue"),

`green  = jmm_cols("green", "yellow"),

`grey  = jmm_cols("light grey", "dark grey")
)

jmm_pal("cool")(5)
}
}
