% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_jmm.R
\name{scale_color_jmm}
\alias{scale_color_jmm}
\title{Construction of the jmm color scales (scale_color_jmm)}
\usage{
scale_color_jmm(palette = "main", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{palette}{the corporate color palette to use}

\item{reverse}{should the color palette be dispayed in a reverse order?}
}
\description{
This function constructs the jmm color scales
}
\examples{
\dontrun{

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
    geom_point(size = 4) +
    scale_color_jmm()

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
    geom_point(size = 4, alpha = .6) +
    scale_color_jmm(discrete = FALSE, palette = "grey")

}
}
