% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_jmm.R
\name{scale_fill_jmm}
\alias{scale_fill_jmm}
\title{Construction of the jmm color fill scales (scale_fill_jmm)}
\usage{
scale_fill_jmm(palette = "main", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{palette}{the corporate color palette to use}

\item{reverse}{should the color palette be dispayed in a reverse order?}
}
\description{
This function constructs the jmm color fill scales
}
\examples{
\dontrun{

ggplot(mpg, aes(manufacturer, fill = manufacturer)) +
    geom_bar() +
    theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
    scale_fill_jmm(palette = "cool", guide = "none")

}
}
